+++ app/renderer/pointer.js
@@ -26,7 +26,17 @@ const {
     getMouseConf,
     listPages
 } = require("./common")
-const {matchesQuery, propPixels, sendToPageOrSubFrame} = require("../util")
+const {
+    matchesQuery,
+    propPixels,
+    sendToPageOrSubFrame,
+    appData,
+    joinPath,
+    readJSON,
+    writeJSON,
+    urlToString,
+    domainName
+} = require("../util")
 
 let X = 0
 let Y = 0
@@ -181,9 +191,9 @@ const storeMouseSelection = selection => {
 
 // ACTIONS
 
-const start = (customX = null, customY = null) => {
-    X = customX || Number(currentPage().getAttribute("pointer-x")) || X
-    Y = customY || Number(currentPage().getAttribute("pointer-y")) || Y
+const start = args => {
+    X = args?.x || Number(currentPage().getAttribute("pointer-x")) || X
+    Y = args?.y || Number(currentPage().getAttribute("pointer-y")) || Y
     const {setMode} = require("./modes")
     setMode("pointer")
     sendToPageOrSubFrame("send-input-event", {"type": "hover", "x": X, "y": Y})
@@ -205,7 +215,10 @@ const moveToMouse = () => {
                     Y = mousePos.y - pagePos.top
                     updateElement()
                 } else {
-                    start(mousePos.x - pagePos.left, mousePos.y - pagePos.top)
+                    start({
+                        "x": mousePos.x - pagePos.left,
+                        "y": mousePos.y - pagePos.top
+                    })
                 }
             }
         })
@@ -600,6 +613,71 @@ const moveFastUp = () => {
     updateElement()
 }
 
+const storePos = args => {
+    const key = args?.key ?? args?.actionCallKeys?.at(-1)
+    if (!key) {
+        return
+    }
+    let posType = getSetting("pointerpostype")
+    if (posType !== "local" && posType !== "global") {
+        posType = "global"
+        if (key !== key.toUpperCase()) {
+            posType = "local"
+        }
+    }
+    const qm = readJSON(joinPath(appData(), "quickmarks")) ?? {}
+    if (!qm.pointer) {
+        qm.pointer = {"global": {}, "local": {}}
+    }
+    if (args?.path === "global") {
+        posType = "global"
+    }
+    if (posType === "local") {
+        let path = ""
+        const pointerPosId = getSetting("pointerposlocalid")
+        if (pointerPosId === "domain") {
+            path = domainName(urlToString(currentPage().src))
+                || domainName(currentPage().src)
+        }
+        if (pointerPosId === "url" || !path) {
+            path = urlToString(currentPage().src) || currentPage().src
+        }
+        path = args?.path ?? path
+        if (!qm.pointer.local[path]) {
+            qm.pointer.local[path] = {}
+        }
+        qm.pointer.local[path][key] = args?.location
+            ?? {"x": Math.round(X), "y": Math.round(Y)}
+    } else {
+        qm.pointer.global[key] = args?.location
+            ?? {"x": Math.round(X), "y": Math.round(Y)}
+    }
+    writeJSON(joinPath(appData(), "quickmarks"), qm)
+}
+
+const restorePos = args => {
+    const key = args?.key ?? args?.actionCallKeys?.at(-1)
+    if (!key) {
+        return
+    }
+
+    const pointerPosId = getSetting("pointerposlocalid")
+    let path = ""
+    if (pointerPosId === "domain") {
+        path = domainName(urlToString(currentPage().src))
+            || domainName(currentPage().src)
+    }
+    if (pointerPosId === "url" || !path) {
+        path = urlToString(currentPage().src) || currentPage().src
+    }
+    path = args?.path ?? path
+    const qm = readJSON(joinPath(appData(), "quickmarks"))
+    const pos = qm?.pointer?.local?.[path]?.[key] ?? qm?.pointer?.global?.[key]
+    if (pos) {
+        move(pos.x, pos.y)
+    }
+}
+
 module.exports = {
     centerOfView,
     copyAudio,
@@ -659,6 +737,7 @@ module.exports = {
     openText,
     openVideo,
     releaseKeys,
+    restorePos,
     restoreSelection,
     rightClick,
     scrollDown,
@@ -677,6 +756,7 @@ module.exports = {
     startOfView,
     startVisualSelect,
     storeMouseSelection,
+    storePos,
     swapPosition,
     toggleMediaControls,
     toggleMediaLoop,